♻️ 5H7BAA code: split cloud push helper
